Save Image modal
- Modal save image, error because parse between int and NaN
- New images for save modal

+++ EditScatterPane.py
@@ -1,6 +1,6 @@
 import tkinter as tk
 
-from tkinter import Canvas, Button, PhotoImage, Entry, StringVar, IntVar, ttk, Checkbutton, Frame, BOTH
+from tkinter import Canvas, Button, DoubleVar, PhotoImage, Entry, StringVar, IntVar, ttk, Checkbutton, Frame, BOTH
 from pathlib import Path
 from text import TEXT, TREETYPE_TIDs_PIDs
 from checkBoxTreeview import CheckboxTreeview
@@ -154,7 +154,7 @@ class ScatterPane(tk.Frame):
             font=("Inter", 15 * -1)
         )
 
-        self.minOutlier_entry = IntVar()
+        self.minOutlier_entry = DoubleVar()
 
         self.entry_image_1 = PhotoImage(
             file=relative_to_assets("entry_2.png"))
@@ -177,7 +177,7 @@ class ScatterPane(tk.Frame):
             height=33.0
         )
 
-        self.maxOutlier_entry = IntVar()
+        self.maxOutlier_entry = DoubleVar()
 
         self.entry_image_2 = PhotoImage(
             file=relative_to_assets("entry_2.png"))
